requirements.txt
torch==2.3.0
torchvision==0.18.0
omegaconf==2.3.0
hydra-core==1.3.2
webdataset==0.2.86
vrs==1.2.1
fsspec==2024.6.0
einops==0.8.0
pandas==2.2.2
pyquaternion==0.9.9
opencv-python==4.10.0.84
tqdm==4.66.4
matplotlib==3.9.0
numpy==1.26.4
moderngl==5.8.2
trimesh==4.4.9
scikit-image==0.24.0
projectaria_tools
